# ice-bubble-admin/src/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ice-bubble Admin API
  description: ice-bubble 管理后台 REST API 文档
  version: 1.4.1
  contact:
    name: ice-bubble team
servers:
  - url: http://localhost:13000
    description: 本地开发
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer Token 认证（配置在 auth.token 或 ICE_AUTH_TOKEN）
  schemas:
...
    Session:
      type: object
      properties:
        session_key:
          type: string
        agent_id:
          type: string
          nullable: true
        agent_name:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        channel:
          type: string
          nullable: true
        message_count:
          type: integer
        first_message:
          type: string
          nullable: true
        last_message:
          type: string
          nullable: true
        last_message_at:
          type: string
          format: date-time
          nullable: true
        first_message_at:
          type: string
          format: date-time
          nullable: true
        session_status:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        model_provider:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
...
paths:
...
  /api/sessions:
    get:
      tags: [Sessions]
      summary: 会话列表（Admin SQLite）
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: agent_id
          in: query
          schema: { type: string }
        - name: channel
          in: query
          schema: { type: string }
        - name: platform
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  total:
                    type: integer
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
...
  /api/messages:
    get:
      tags: [Messages]
      summary: 消息列表
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: session_key
          in: query
          schema: { type: string }
        - name: archived
          in: query
          description: "true 查询归档数据"
          schema: { type: string }
      responses:
        '200':
          description: 消息列表
...